refactor(epub-preview): extract parse and toc modules
